Incorporate timeout for test cases
Adjust the TestCase functionality to accommodate TimeoutProcess. A test
case can have a class level timeout applied. Add a new method to create
a process within a test, which will automatically apply the timeout if
it is configured. If a timeout occurs, mark the test as failed due to
a timeout.

git-svn-id: https://svn.geocomp.uq.edu.au/svn/jeathesis@39 a4bfb809-27fc-4812-b90d-531e79a1b4ec