Marks: Add Verbose test mode
Allow a single test to be run in verbose mode. This will show a diff
of the stderr or stdout output that lead to a test failure.
If no test is specified, verbose mode is ignored.
If save mode is used with verbose mode, show a full diff of the output.
Currently only works with output comparison with a file.

git-svn-id: https://svn.geocomp.uq.edu.au/svn/jeathesis@107 a4bfb809-27fc-4812-b90d-531e79a1b4ec