Marks: Fix multiple marking mode - directory path and output files
Fix a bug where the only working directory was '.'. The path specified
is now properly used, meaning any path can be specified to run tests.
Also allow the output file paths to be specified as options in the
marking script.
Sort the CSV output by submission ID (folder name), so output is consistent.

git-svn-id: https://svn.geocomp.uq.edu.au/svn/jeathesis@130 a4bfb809-27fc-4812-b90d-531e79a1b4ec