Marks: Fix issue with closed streams after timeout
Change all methods using streams to have a pointer to the stream. This
allows the streams to be checked for closure before attempting to read
from them.
This was an issue with TimeoutProcess, as the timeout would close streams
from another thread underneath the main thread trying to read input.

git-svn-id: https://svn.geocomp.uq.edu.au/svn/jeathesis@131 a4bfb809-27fc-4812-b90d-531e79a1b4ec